git/code/requirements.txt
streamlit
pandas
scrapy
openpyxl
crochet
google.generativeai
